move extract packet to lib
